warning fix

+++ source/rcnn/modeling/roi_heads/box_head/roi_box_feature_extractors.cpp
@@ -4,12 +4,12 @@
 #include <cmath>
 
 
-namespace rcnn{
-namespace modeling{
+namespace rcnn {
+namespace modeling {
 
 ResNet50Conv5ROIFeatureExtractorImpl::ResNet50Conv5ROIFeatureExtractorImpl(int64_t in_channels)
   :pooler_(register_module("pooler", MakePooler("ROI_BOX_HEAD"))),
-  head_(register_module("head",
+   head_(register_module("head",
     ResNetHead(
       std::vector<ResNetImpl::StageSpec>{ResNetImpl::StageSpec(4, 3, false)},
       rcnn::config::GetCFG<int64_t>({"MODEL", "RESNETS", "NUM_GROUPS"}),
@@ -20,7 +20,7 @@ ResNet50Conv5ROIFeatureExtractorImpl::ResNet50Conv5ROIFeatureExtractorImpl(int64
       rcnn::config::GetCFG<int64_t>({"MODEL", "RESNETS", "RES5_DILATION"})
     )
   )),
-  out_channels_(head_->out_channels()){}
+  out_channels_(head_->out_channels()) {}
 
 torch::Tensor ResNet50Conv5ROIFeatureExtractorImpl::forward(std::vector<torch::Tensor> x, std::vector<rcnn::structures::BoxList> proposals){
   torch::Tensor output = pooler_->forward(x, proposals);
@@ -52,16 +52,16 @@ torch::Tensor FPN2MLPFeatureExtractorImpl::forward(std::vector<torch::Tensor> x,
   return output;
 }
 
-int64_t FPN2MLPFeatureExtractorImpl::out_channels() const{
+int64_t FPN2MLPFeatureExtractorImpl::out_channels() const {
   return out_channels_;
 }
 
-torch::nn::Sequential FPNXconv1fcFeatureExtractorImpl::make_xconvs(int64_t in_channels){
+torch::nn::Sequential FPNXconv1fcFeatureExtractorImpl::make_xconvs(int64_t in_channels) {
   torch::nn::Sequential xconvs;
   int64_t conv_head_dim = rcnn::config::GetCFG<int64_t>({"MODEL", "ROI_BOX_HEAD", "CONV_HEAD_DIM"});
   int64_t num_stacked_convs = rcnn::config::GetCFG<int64_t>({"MODEL", "ROI_BOX_HEAD", "NUM_STACKED_CONVS"});
   int64_t dilation = rcnn::config::GetCFG<int64_t>({"MODEL", "ROI_BOX_HEAD", "DILATION"});
-  for(size_t i = 0; i < num_stacked_convs; ++i){
+  for (int64_t i = 0; i < num_stacked_convs; ++i) {
     xconvs->push_back(
       torch::nn::Conv2d(
         torch::nn::Conv2dOptions(in_channels, conv_head_dim, 3).stride(1).padding(dilation).dilation(dilation).with_bias(true)
@@ -70,11 +70,11 @@ torch::nn::Sequential FPNXconv1fcFeatureExtractorImpl::make_xconvs(int64_t in_ch
     in_channels = conv_head_dim;
     xconvs->push_back(torch::nn::Functional(torch::relu));
   }
-  for(auto& param : xconvs->named_parameters()){
-    if(param.key().find("weight") != std::string::npos) {
+  for (auto& param : xconvs->named_parameters()) {
+    if (param.key().find("weight") != std::string::npos) {
       torch::nn::init::normal_(param.value(), 0.01);
     }
-    else if(param.key().find("bias") != std::string::npos) {
+    else if (param.key().find("bias") != std::string::npos) {
       torch::nn::init::constant_(param.value(), 0);
     }
   }
@@ -83,12 +83,12 @@ torch::nn::Sequential FPNXconv1fcFeatureExtractorImpl::make_xconvs(int64_t in_ch
 
 FPNXconv1fcFeatureExtractorImpl::FPNXconv1fcFeatureExtractorImpl(int64_t in_channels)
   :pooler_(register_module("pooler", MakePooler("ROI_BOX_HEAD"))),
-  xconvs_(register_module("xconvx", make_xconvs(in_channels))),
-  fc6_(register_module("fc6", layers::MakeFC(
+   xconvs_(register_module("xconvx", make_xconvs(in_channels))),
+   fc6_(register_module("fc6", layers::MakeFC(
       rcnn::config::GetCFG<int64_t>({"MODEL", "ROI_BOX_HEAD", "CONV_HEAD_DIM"}) * pow(rcnn::config::GetCFG<int64_t>({"MODEL", "ROI_BOX_HEAD", "POOLER_RESOLUTION"}), 2),
       rcnn::config::GetCFG<int64_t>({"MODEL", "ROI_BOX_HEAD", "MLP_HEAD_DIM"})
     ))),
-  out_channels_(rcnn::config::GetCFG<int64_t>({"MODEL", "ROI_BOX_HEAD", "MLP_HEAD_DIM"})){}
+  out_channels_(rcnn::config::GetCFG<int64_t>({"MODEL", "ROI_BOX_HEAD", "MLP_HEAD_DIM"})) {}
 
 torch::Tensor FPNXconv1fcFeatureExtractorImpl::forward(std::vector<torch::Tensor> x, std::vector<rcnn::structures::BoxList> proposals){
   torch::Tensor output = pooler_->forward(x, proposals);
@@ -98,34 +98,34 @@ torch::Tensor FPNXconv1fcFeatureExtractorImpl::forward(std::vector<torch::Tensor
   return output;
 }
 
-int64_t FPNXconv1fcFeatureExtractorImpl::out_channels() const{
+int64_t FPNXconv1fcFeatureExtractorImpl::out_channels() const {
   return out_channels_;
 }
 
-std::pair<torch::nn::Sequential, int64_t> MakeROIBoxFeatureExtractor(int64_t in_channels){
+std::pair<torch::nn::Sequential, int64_t> MakeROIBoxFeatureExtractor(int64_t in_channels) {
   torch::nn::Sequential extractor;
   int64_t out_channels;
   std::string name = rcnn::config::GetCFG<std::string>({"MODEL", "ROI_BOX_HEAD", "FEATURE_EXTRACTOR"});
-  if(name.compare("ResNet50Conv5ROIFeatureExtractor") == 0){
+  if (name.compare("ResNet50Conv5ROIFeatureExtractor") == 0) {
     auto model = ResNet50Conv5ROIFeatureExtractor(in_channels);
     extractor->push_back(model);
     out_channels = model->out_channels();
   }
-  else if(name.compare("FPN2MLPFeatureExtractor") == 0){
+  else if (name.compare("FPN2MLPFeatureExtractor") == 0) {
     auto model = FPN2MLPFeatureExtractor(in_channels);
     extractor->push_back(model);
     out_channels = model->out_channels();
   }
-  else if(name.compare("FPNXconv1fcFeatureExtractor") == 0){
+  else if (name.compare("FPNXconv1fcFeatureExtractor") == 0) {
     auto model = FPNXconv1fcFeatureExtractor(in_channels);
     extractor->push_back(model);
     out_channels = model->out_channels();
   }
-  else{
+  else {
     assert(false);
   }
   return std::make_pair(extractor, out_channels);
 }
 
-}
-}
+} // namespace modeling
+} // namespace rcnn
